requirements.txt
requests
python-whois
dnspython
fpdf
urllib3
